requirements-dev.txt
alembic==1.11.3
annotated-types==0.5.0
anyio==3.7.1
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
et-xmlfile==1.1.0
exceptiongroup==1.1.3
fastapi==0.103.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
iniconfig==2.0.0
lxml==4.9.3
Mako==1.2.4
MarkupSafe==2.1.3
openpyxl==3.1.2
packaging==23.1
pluggy==1.3.0
psycopg2-binary==2.9.7
pydantic==2.3.0
pydantic-settings==2.0.3
pydantic_core==2.6.3
pytest==7.4.0
pytest-mock==3.11.1
python-dotenv==1.0.0
requests==2.31.0
requests-toolbelt==1.0.0
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==2.0.20
starlette==0.27.0
tomli==2.0.1
typing_extensions==4.7.1
urllib3==2.0.4
uvicorn==0.23.2
